pyproject.toml
[project]
name = "poizon-seller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.53.1",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = ["-m", "not e2e", "-q"]
markers = [
    "e2e: live Poizon/Musinsa API (credentials and network required)",
]
